Extract the review right side context into a separate file RightSide.jsx.
