+++ src/components/preview/ui/Preview.jsx
@@ -1,17 +1,13 @@
 /* eslint-disable react/jsx-no-undef */
 import {FaFacebook, FaGithub, FaInstagram, FaLinkedin, FaTwitter, FaYoutube,} from "react-icons/fa";
 import {CgWebsite} from "react-icons/cg";
-import Skills from "../components/Skills";
-import DateRange from "../../utility/DateRange";
-import Link from "next/link";
 import React, {useContext, useState} from "react";
 import {ResumeContext} from "../../builder";
 import dynamic from "next/dynamic";
-import Language from "../components/Language";
-import Certification from "../components/Certification";
 import ModalHighlightMenu from "../components/ModalHighlightMenu";
 import Header from "../components/Header";
 import LeftSide from "../components/LeftSide";
+import RightSide from "../components/RightSide";
 
 const DragDropContext = dynamic(
   () =>
@@ -20,24 +16,9 @@ const DragDropContext = dynamic(
     }),
   {ssr: false}
 );
-const Droppable = dynamic(
-  () =>
-    import("react-beautiful-dnd").then((mod) => {
-      return mod.Droppable;
-    }),
-  {ssr: false}
-);
-const Draggable = dynamic(
-  () =>
-    import("react-beautiful-dnd").then((mod) => {
-      return mod.Draggable;
-    }),
-  {ssr: false}
-);
 
 const Preview = () => {
   const {resumeData, setResumeData} = useContext(ResumeContext);
-  const [content, setContent] = useState(resumeData);
   const icons = [
     {name: "github", icon: <FaGithub/>},
     {name: "linkedin", icon: <FaLinkedin/>},
@@ -111,208 +92,9 @@ const Preview = () => {
         <DragDropContext onDragEnd={onDragEnd}>
           <Header resumeData={resumeData} icons={icons}/>
           <hr className="border-dashed my-2"/>
-          {/* two column start */}
           <div className="grid grid-cols-3 gap-6">
-            <LeftSide resumeData={resumeData} />
-            {/* Right side start */}
-            <div className="col-span-2 space-y-2">
-              {resumeData.workExperience.length > 0 && (
-                <Droppable droppableId="work-experience" type="WORK_EXPERIENCE">
-                  {(provided) => (
-                    <div {...provided.droppableProps} ref={provided.innerRef}>
-                      <h2
-                        className="section-title mb-1 border-b-2 border-gray-300 editable"
-                        contentEditable
-                        suppressContentEditableWarning
-                      >
-                        Work Experience
-                      </h2>
-                      {resumeData.workExperience.map((item, index) => (
-                        <Draggable
-                          key={`${item.company}-${index}`}
-                          draggableId={`WORK_EXPERIENCE-${index}`}
-                          index={index}
-                        >
-                          {(provided, snapshot) => (
-                            <div
-                              ref={provided.innerRef}
-                              {...provided.draggableProps}
-                              {...provided.dragHandleProps}
-                              className={`mb-1 ${
-                                snapshot.isDragging &&
-                                "outline-dashed outline-2 outline-gray-400 bg-white"
-                              }`}
-                            >
-                              <div className="flex flex-row justify-between space-y-1">
-                                <p className="content i-bold">{item.company}</p>
-                                <DateRange
-                                  startYear={item.startYear}
-                                  endYear={item.endYear}
-                                  id={`work-experience-start-end-date`}
-                                />
-                              </div>
-                              <p className="content">{item.position}</p>
-                              <p className="content hyphens-auto">
-                                {item.description}
-                              </p>
-                              <Droppable
-                                droppableId={`WORK_EXPERIENCE_KEY_ACHIEVEMENT-${index}`}
-                                type="WORK_EXPERIENCE_KEY_ACHIEVEMENT"
-                              >
-                                {(provided) => (
-                                  <ul
-                                    className="list-disc ul-padding content"
-                                    {...provided.droppableProps}
-                                    ref={provided.innerRef}
-                                  >
-                                    {typeof item.keyAchievements === "string" &&
-                                      item.keyAchievements
-                                        .split("\n")
-                                        .map((achievement, subIndex) => (
-                                          <Draggable
-                                            key={`${item.company}-${index}-${subIndex}`}
-                                            draggableId={`WORK_EXPERIENCE_KEY_ACHIEVEMENT-${index}-${subIndex}`}
-                                            index={subIndex}
-                                          >
-                                            {(provided, snapshot) => (
-                                              <li
-                                                ref={provided.innerRef}
-                                                {...provided.draggableProps}
-                                                {...provided.dragHandleProps}
-                                                className={`
-                                          hover:outline-dashed hover:outline-2 hover:outline-gray-400
-                                          ${
-                                                  snapshot.isDragging &&
-                                                  "outline-dashed outline-2 outline-gray-400 bg-white"
-                                                }`}
-                                              >
-                                                <div
-                                                  dangerouslySetInnerHTML={{
-                                                    __html: achievement,
-                                                  }}
-                                                  contentEditable
-                                                />
-                                              </li>
-                                            )}
-                                          </Draggable>
-                                        ))}
-                                    {provided.placeholder}
-                                  </ul>
-                                )}
-                              </Droppable>
-                            </div>
-                          )}
-                        </Draggable>
-                      ))}
-                      {provided.placeholder}
-                    </div>
-                  )}
-                </Droppable>
-              )}
-              {resumeData.projects.length > 0 && (
-                <Droppable droppableId="projects" type="PROJECTS">
-                  {(provided) => (
-                    <div {...provided.droppableProps} ref={provided.innerRef}>
-                      <h2
-                        className="section-title mb-1 border-b-2 border-gray-300 editable"
-                        contentEditable
-                        suppressContentEditableWarning
-                      >
-                        Projects
-                      </h2>
-                      {resumeData.projects.map((item, index) => (
-                        <Draggable
-                          key={`${item.name}-${index}`}
-                          draggableId={`PROJECTS-${index}`}
-                          index={index}
-                        >
-                          {(provided, snapshot) => (
-                            <div
-                              ref={provided.innerRef}
-                              {...provided.draggableProps}
-                              {...provided.dragHandleProps}
-                              className={`mb-1 ${
-                                snapshot.isDragging &&
-                                "outline-dashed outline-2 outline-gray-400 bg-white"
-                              }`}
-                            >
-                              <div className="flex flex-row justify-between space-y-1">
-                                <p className="content i-bold">{item.name}</p>
-                                <DateRange
-                                  startYear={item.startYear}
-                                  endYear={item.endYear}
-                                  id={`work-experience-start-end-date`}
-                                />
-                              </div>
-
-                              <Link
-                                href={item.link}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="content"
-                              >
-                                {item.link}
-                              </Link>
-                              <p className="content">{item.description}</p>
-                              <Droppable
-                                droppableId={`PROJECTS_KEY_ACHIEVEMENT-${index}`}
-                                type="PROJECTS_KEY_ACHIEVEMENT"
-                              >
-                                {(provided) => (
-                                  <ul
-                                    className="list-disc ul-padding content"
-                                    {...provided.droppableProps}
-                                    ref={provided.innerRef}
-                                  >
-                                    {typeof item.keyAchievements === "string" &&
-                                      item.keyAchievements
-                                        .split("\n")
-                                        .map((achievement, subIndex) => (
-                                          <Draggable
-                                            key={`${item.name}-${index}-${subIndex}`}
-                                            draggableId={`PROJECTS_KEY_ACHIEVEMENT-${index}-${subIndex}`}
-                                            index={subIndex}
-                                          >
-                                            {(provided, snapshot) => (
-                                              <li
-                                                ref={provided.innerRef}
-                                                {...provided.draggableProps}
-                                                {...provided.dragHandleProps}
-                                                className={`
-                                          hover:outline-dashed hover:outline-2 hover:outline-gray-400
-                                          ${
-                                                  snapshot.isDragging &&
-                                                  "outline-dashed outline-2 outline-gray-400 bg-white"
-                                                }`}
-                                              >
-                                                <div
-                                                  dangerouslySetInnerHTML={{
-                                                    __html: achievement,
-                                                  }}
-                                                  contentEditable
-                                                />
-                                              </li>
-                                            )}
-                                          </Draggable>
-                                        ))}
-                                    {provided.placeholder}
-                                  </ul>
-                                )}
-                              </Droppable>
-                            </div>
-
-                          )}
-
-                        </Draggable>
-                      ))}
-                      {provided.placeholder}
-                    </div>
-                  )}
-                </Droppable>
-              )}
-            </div>
-            {/* Rigth side end */}
-
+            <LeftSide resumeData={resumeData}/>
+            <RightSide resumeData={resumeData}/>
           </div>
         </DragDropContext>
       </A4PageWrapper>
